+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # Unreleased
 
+*   Add a `kubernetes` section in `shipit.yml` for first class k8s support.
+
 *   Disregard GitHub's Cache-Control max-age directives, because they impose a 60 seconds resolution
     which is way too slow.